requirements.txt
duckduckgo-search==3.9.8
langchain==0.0.352
matplotlib>=3.8.0
openai>=0.27.7
openpyxl>=3.1.2
pandas>=2.1.1
python-box>=7.1.1
python-dotenv>=1.0.0
xlsxwriter>=3.1.9
wikipedia>=1.4.0
